requirements.txt
appdirs==1.4.4
Babel==2.9.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
Brotli==1.0.9
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cssselect==1.1.0
dash==1.20.0
dash-bootstrap-components==0.12.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
feedparser==6.0.2
Flask==1.1.2
Flask-Compress==1.9.0
forex-python==1.5
future==0.18.2
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
paramiko==2.7.2
parse==1.19.0
Pillow==8.2.0
plotly==4.14.3
psycopg2==2.8.6
pycparser==2.20
pyee==8.1.0
PyNaCl==1.4.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-html==0.10.0
retrying==1.3.3
sgmllib3k==1.0.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.9
sshtunnel==0.4.0
tqdm==4.60.0
urllib3==1.26.4
w3lib==1.22.0
websockets==8.1
Werkzeug==1.0.1
yahoo-fin==0.8.8
